pyproject.toml
[tool.poetry]
name = "cloud-gm"
version = "0.1.0"
description = ""
authors = ["Ryan Batchelder <ryanbatch@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
python-fasthtml = "^0.9.0"
boto3 = "^1.35.54"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
